+++ R/utils_rt.R
@@ -86,8 +86,7 @@ outro <- function(
     if (real_user_system) {
       msg0(
         paste0(
-          checkmark(),
-          " Done in ",
+          "Done in ",
           format_seconds(elapsed[3]),
           " (",
           "Real:",
@@ -103,8 +102,7 @@ outro <- function(
     } else {
       msg0(
         paste0(
-          checkmark(),
-          " Done in ",
+          "Done in ",
           format_seconds(elapsed[3]),
           "."
         ),